.windsurf/skills/dtx/dtx-validate/scripts/include/pyproject.toml
[project]
name = "dtx-validate-cli"
version = "0.1.0"
description = "CLI for dtx-validate skill"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
